Fix Audio is not muted.
:Release Notes:
Fix Audio is not muted in MCS test application.

:Detailed Notes:
In MCS there is condition to send the mediaId, It will expect the
particular displayId to send the mediaId.
DisplayId is extracted from appId. But in case of WebosOSE upper layer
is sending only appId, It's not append with displayId. So as of now MCS
hard coding the displayId as zero for WebosOSE.

:Testing Performed:
QWA-4233, QWA-4236, QWA-4237, QWA-5391

:QA Notes:
NA

:Issues Addressed:
[PLAT-130572] [webOS OSE]: Audio is not muted.

Change-Id: I865c0be48a4cc9e8f3d950ba452ebe7db84f4174
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.mediacontroller/+/300181
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: <siva.kannappan@lge.com>
Reviewed-by: <manisha.manisha@lge.com>
Reviewed-by: <kalaimani.k@lge.com>
Tested-by: <siva.kannappan@lge.com>